examples/local/news-agent/agents/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.34.0
uvicorn>=0.30.0
fastapi>=0.111.0
mcp>=1.0.0
feedparser>=6.0.0
